requirements.txt
django-storages
boto
pillow
